HO5.0/requirements.txt
Flask>=2.0
scikit-learn>=1.0
pandas>=1.3
PyPDF2>=3.0
python-docx>=0.8
